feat(#53): 타이머 hook, component 분리
